api/requirements.txt
fastapi==0.65.2
uvicorn==0.13.4
typing == 3.7.4.3
pydantic==1.8.2
requests==2.27.1
python-keycloak==0.24.0
fastapi-cprofile==0.0.2 
requests-async==0.6.2